requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.2.2
WTForms==3.2.1
Werkzeug==2.3.7
python-dotenv==1.0.0
